# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Aaditya Fabrication Works API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export type InquiryStatus = typeof InquiryStatus[keyof typeof InquiryStatus];


export const InquiryStatus = {
  unread: 'unread',
  read: 'read',
} as const;

export interface Inquiry {
  id: number;
  name: string;
  /** @nullable */
  companyName?: string | null;
  phone: string;
  email: string;
  service: string;
  /** @nullable */
  message?: string | null;
  /** @nullable */
  requiredCapacity?: number | null;
  /** @nullable */
  spanMeters?: number | null;
  /** @nullable */
  shedDimensions?: string | null;
  /** @nullable */
  existingEquipment?: string | null;
  /** @nullable */
  adminNotes?: string | null;
  status: InquiryStatus;
  createdAt: string;
}

export interface InquiryList {
  data: Inquiry[];
  total: number;
  page: number;
  limit: number;
}

export interface InquiryInput {
  /** @minLength 2 */
  name: string;
  companyName?: string;
  phone: string;
  email: string;
  service: string;
  message?: string;
  requiredCapacity?: number;
  spanMeters?: number;
  shedDimensions?: string;
  existingEquipment?: string;
}

export type InquiryUpdateStatus = typeof InquiryUpdateStatus[keyof typeof InquiryUpdateStatus];


export const InquiryUpdateStatus = {
  unread: 'unread',
  read: 'read',
} as const;

export interface InquiryUpdate {
  status?: InquiryUpdateStatus;
  adminNotes?: string;
}

export type InquiryStatsByServiceItem = {
  service: string;
  count: number;
};

export interface InquiryStats {
  total: number;
  unread: number;
  recentCount: number;
  byService: InquiryStatsByServiceItem[];
}

export interface Project {
  id: number;
  title: string;
  category: string;
  description: string;
  /** @nullable */
  beforeImage?: string | null;
  /** @nullable */
  afterImage?: string | null;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  year?: number | null;
  /** @nullable */
  specs?: string | null;
  featured: boolean;
  createdAt: string;
}

export interface ProjectInput {
  title: string;
  category: string;
  description: string;
  beforeImage?: string;
  afterImage?: string;
  location?: string;
  year?: number;
  specs?: string;
  featured?: boolean;
}

export interface ProjectUpdate {
  title?: string;
  category?: string;
  description?: string;
  beforeImage?: string;
  afterImage?: string;
  location?: string;
  year?: number;
  specs?: string;
  featured?: boolean;
}

export interface Service {
  id: number;
  name: string;
  slug: string;
  shortDescription: string;
  fullDescription: string;
  icon: string;
  /** @nullable */
  specs?: string | null;
  sortOrder: number;
}

export interface ServiceInput {
  name: string;
  slug: string;
  shortDescription: string;
  fullDescription: string;
  icon: string;
  specs?: string;
  sortOrder?: number;
}

export interface ServiceUpdate {
  name?: string;
  slug?: string;
  shortDescription?: string;
  fullDescription?: string;
  icon?: string;
  specs?: string;
  sortOrder?: number;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface AdminSession {
  authenticated: boolean;
  email: string;
}

export type ListInquiriesParams = {
status?: ListInquiriesStatus;
service?: string;
search?: string;
page?: number;
limit?: number;
};

export type ListInquiriesStatus = typeof ListInquiriesStatus[keyof typeof ListInquiriesStatus];


export const ListInquiriesStatus = {
  all: 'all',
  read: 'read',
  unread: 'unread',
} as const;

export type ListProjectsParams = {
category?: string;
featured?: boolean;
};