# lib/api-zod/src/generated/types/inquiryInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Aaditya Fabrication Works API
 * OpenAPI spec version: 0.1.0
 */

export interface InquiryInput {
  /** @minLength 2 */
  name: string;
  companyName?: string;
  phone: string;
  email: string;
  service: string;
  message?: string;
  requiredCapacity?: number;
  spanMeters?: number;
  shedDimensions?: string;
  existingEquipment?: string;
}