# lib/api-zod/src/generated/types/inquiryList.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Aaditya Fabrication Works API
 * OpenAPI spec version: 0.1.0
 */
import type { Inquiry } from './inquiry';

export interface InquiryList {
  data: Inquiry[];
  total: number;
  page: number;
  limit: number;
}