# lib/api-zod/src/generated/types/inquiryStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Aaditya Fabrication Works API
 * OpenAPI spec version: 0.1.0
 */
import type { InquiryStatsByServiceItem } from './inquiryStatsByServiceItem';

export interface InquiryStats {
  total: number;
  unread: number;
  recentCount: number;
  byService: InquiryStatsByServiceItem[];
}