# lib/api-zod/src/generated/types/projectUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Aaditya Fabrication Works API
 * OpenAPI spec version: 0.1.0
 */

export interface ProjectUpdate {
  title?: string;
  category?: string;
  description?: string;
  beforeImage?: string;
  afterImage?: string;
  location?: string;
  year?: number;
  specs?: string;
  featured?: boolean;
}