# lib/api-zod/src/generated/types/service.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Aaditya Fabrication Works API
 * OpenAPI spec version: 0.1.0
 */

export interface Service {
  id: number;
  name: string;
  slug: string;
  shortDescription: string;
  fullDescription: string;
  icon: string;
  /** @nullable */
  specs?: string | null;
  sortOrder: number;
}