# lib/api-zod/src/generated/types/serviceUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Aaditya Fabrication Works API
 * OpenAPI spec version: 0.1.0
 */

export interface ServiceUpdate {
  name?: string;
  slug?: string;
  shortDescription?: string;
  fullDescription?: string;
  icon?: string;
  specs?: string;
  sortOrder?: number;
}